Fixed compile warnings for size_t (Michael Friedrich)
git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1730 5f96b256-904b-4d8d-8c98-d829582c6739

+++ common/macros.c
@@ -146,7 +146,7 @@ int process_macros_r(nagios_macros *mac, char *input_buffer, char **output_buffe
 			*output_buffer=(char *)realloc(*output_buffer,strlen(*output_buffer)+strlen(temp_buffer)+1);
 			strcat(*output_buffer,temp_buffer);
 
-			log_debug_info(DEBUGL_MACROS,2,"  Not currently in macro.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+			log_debug_info(DEBUGL_MACROS,2,"  Not currently in macro.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 			in_macro=TRUE;
 			}
 
@@ -173,14 +173,14 @@ int process_macros_r(nagios_macros *mac, char *input_buffer, char **output_buffe
 
 			/* an escaped $ is done by specifying two $$ next to each other */
 			else if(!strcmp(temp_buffer,"")){
-				log_debug_info(DEBUGL_MACROS,2,"  Escaped $.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+				log_debug_info(DEBUGL_MACROS,2,"  Escaped $.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 				*output_buffer=(char *)realloc(*output_buffer,strlen(*output_buffer)+2);
 				strcat(*output_buffer,"$");
 				}
 
 			/* a non-macro, just some user-defined string between two $s */
 			else{
-				log_debug_info(DEBUGL_MACROS,2,"  Non-macro.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+				log_debug_info(DEBUGL_MACROS,2,"  Non-macro.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 
 				/* add the plain text to the end of the already processed buffer */
 				*output_buffer=(char *)realloc(*output_buffer,strlen(*output_buffer)+strlen(temp_buffer)+3);
@@ -216,7 +216,7 @@ int process_macros_r(nagios_macros *mac, char *input_buffer, char **output_buffe
 						*output_buffer=(char *)realloc(*output_buffer,strlen(*output_buffer)+strlen(cleaned_macro)+1);
 						strcat(*output_buffer,cleaned_macro);
 
-						log_debug_info(DEBUGL_MACROS,2,"  Cleaned macro.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+						log_debug_info(DEBUGL_MACROS,2,"  Cleaned macro.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 						}
 					}
 
@@ -227,7 +227,7 @@ int process_macros_r(nagios_macros *mac, char *input_buffer, char **output_buffe
 						*output_buffer=(char *)realloc(*output_buffer,strlen(*output_buffer)+strlen(selected_macro)+1);
 						strcat(*output_buffer,selected_macro);
 
-						log_debug_info(DEBUGL_MACROS,2,"  Uncleaned macro.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+						log_debug_info(DEBUGL_MACROS,2,"  Uncleaned macro.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 						}
 					}
 
@@ -235,7 +235,7 @@ int process_macros_r(nagios_macros *mac, char *input_buffer, char **output_buffe
 				if(free_macro==TRUE)
 					my_free(selected_macro);
 
-				log_debug_info(DEBUGL_MACROS,2,"  Just finished macro.  Running output (%d): '%s'\n",strlen(*output_buffer),*output_buffer);
+				log_debug_info(DEBUGL_MACROS,2,"  Just finished macro.  Running output (%lu): '%s'\n",(unsigned long)strlen(*output_buffer),*output_buffer);
 				}
 
 			in_macro=FALSE;
